@@ -4,32 +4,32 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   deriv_chart:
     path: ../
-  vibration: ^1.7.4-nullsafety.0
-  intl: ^0.17.0
+  vibration: ^1.7.6
+  intl: ^0.18.0
   flutter_localizations:
     sdk: flutter
-  pref: ^2.2.0
+  pref: ^2.7.1
   flutter_deriv_api:
     git:
       url: git@github.com:regentmarkets/flutter-deriv-api.git
-      ref: master
+      ref: flutter-version-3
 
 dev_dependencies:
   deriv_lint:
     git:
       url: git@github.com:regentmarkets/flutter-deriv-packages.git
       path: packages/deriv_lint
-      ref: dev
+      ref: flutter-version-3
   flutter_test:
     sdk: flutter
-  intl_utils: ^2.0.0
+  intl_utils: ^2.8.2
 
 flutter:
   uses-material-design: true
