@@ -4,6 +4,7 @@ import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_too
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/drawing_paint_style.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/drawing_parts.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/drawing_pattern.dart';
+import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/edge_point.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/point.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_data.dart';
@@ -16,16 +17,14 @@ import 'package:deriv_chart/deriv_chart.dart';
 /// that is infinite in both directions.
 class VerticalDrawing extends Drawing {
   /// Initializes
-  VerticalDrawing({
-    required this.drawingPart,
-    required this.epochFromX,
-    required this.chartConfig,
-    this.epoch = 0,
-    this.yCoord = 0,
-  });
+  VerticalDrawing(
+      {required this.drawingPart,
+      required this.chartConfig,
+      required this.edgePoint,
+      required this.epochFromX});
 
   /// Chart config to get pipSize
-  final ChartConfig chartConfig;
+  final ChartConfig? chartConfig;
 
   /// Get epoch from x.
   int Function(double x)? epochFromX;
@@ -33,11 +32,8 @@ class VerticalDrawing extends Drawing {
   /// Part of a drawing: 'vertical'
   final DrawingParts drawingPart;
 
-  /// Starting epoch.
-  final int epoch;
-
-  /// Starting Y coordinates.
-  final double yCoord;
+  /// Starting point of drawing
+  final EdgePoint edgePoint;
 
   /// Keeps the latest position of the start of drawing
   Point? startPoint;
@@ -51,6 +47,10 @@ class VerticalDrawing extends Drawing {
     double Function(int x) epochToX,
     double Function(double y) quoteToY,
     DrawingData drawingData,
+    Point Function(
+      EdgePoint edgePoint,
+      DraggableEdgePoint draggableEdgePoint,
+    ) updatePositionCallback,
     DraggableEdgePoint draggableStartPoint, {
     DraggableEdgePoint? draggableEndPoint,
   }) {
@@ -61,12 +61,7 @@ class VerticalDrawing extends Drawing {
     final LineStyle lineStyle = config.lineStyle;
     final DrawingPatterns pattern = config.pattern;
 
-    startPoint = draggableStartPoint.updatePosition(
-      epoch,
-      yCoord,
-      epochToX,
-      quoteToY,
-    );
+    startPoint = updatePositionCallback(edgePoint, draggableStartPoint);
 
     final double xCoord = startPoint!.x;
     final double startQuoteToY = startPoint!.y;
@@ -88,7 +83,7 @@ class VerticalDrawing extends Drawing {
           xCoord,
           'vertical',
           theme,
-          chartConfig,
+          chartConfig!,
           epochFromX: epochFromX,
         );
       }
@@ -103,8 +98,10 @@ class VerticalDrawing extends Drawing {
     double Function(int x) epochToX,
     double Function(double y) quoteToY,
     DrawingToolConfig config,
-    DraggableEdgePoint draggableStartPoint, {
+    DraggableEdgePoint draggableStartPoint,
+    void Function({required bool isDragged}) setIsStartPointDragged, {
     DraggableEdgePoint? draggableEndPoint,
+    void Function({required bool isDragged})? setIsEndPointDragged,
   }) {
     final LineStyle lineStyle = config.toJson()['lineStyle'];
 
