@@ -1,5 +1,6 @@
 import 'dart:ui';
 
+import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/edge_point.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/point.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/x_axis/x_axis_model.dart';
 
@@ -8,44 +9,71 @@ import 'package:deriv_chart/src/deriv_chart/chart/x_axis/x_axis_model.dart';
 /// And we want to hanle difftent types of drag events on them.
 /// For example with dots are draggable edge points for the line
 /// ⎯⎯⚪️⎯⎯⎯⚪️⎯⎯
-class DraggableEdgePoint {
+class DraggableEdgePoint extends EdgePoint {
+  /// Initializes
+  DraggableEdgePoint({
+    int epoch = 0,
+    double quote = 0,
+    this.isDrawingDragged = false,
+    this.isDragged = false,
+  }) : super(epoch: epoch, quote: quote);
+
   /// Represents whether the whole drawing is currently being dragged or not
-  bool isDrawingDragged = false;
+  final bool isDrawingDragged;
 
   /// Represents whether the edge point is currently being dragged or not
-  bool isDragged = false;
+  final bool isDragged;
 
   /// Holds the current position of the edge point when it is being dragged.
-  Offset draggedPosition = Offset.zero;
+  Offset _draggedPosition = Offset.zero;
 
   /// A callback method that takes the relative x and y positions as parameter,
   /// sets the draggedPosition field to its value and return epoch and quote
   /// values.
-  Point updatePosition(int epoch, double yCoord,
-      double Function(int x) epochToX, double Function(double y) quoteToY) {
+  Point updatePosition(
+    int epoch,
+    double yCoord,
+    double Function(int x) epochToX,
+    double Function(double y) quoteToY,
+  ) {
     final Offset oldPosition = Offset(epoch.toDouble(), yCoord);
-    draggedPosition = isDrawingDragged ? draggedPosition : oldPosition;
+    _draggedPosition = isDrawingDragged ? _draggedPosition : oldPosition;
 
-    final double xCoord = epochToX(draggedPosition.dx.toInt());
-    final double quoteY = quoteToY(draggedPosition.dy);
+    final double x = epochToX(_draggedPosition.dx.toInt());
+    final double y = quoteToY(_draggedPosition.dy);
 
-    return Point(x: xCoord, y: quoteY);
+    return Point(x: x, y: y);
   }
 
   /// A method that takes the gesture delta Offset object as parameter
   /// and sets the draggedPosition field to its value.
   void updatePositionWithLocalPositions(
-      Offset delta,
-      XAxisModel xAxis,
-      double Function(double) quoteFromCanvasY,
-      double Function(double y) quoteToY,
-      {required bool isOtherEndDragged}) {
+    Offset delta,
+    XAxisModel xAxis,
+    double Function(double) quoteFromCanvasY,
+    double Function(double y) quoteToY, {
+    required bool isOtherEndDragged,
+  }) {
     final Offset localPosition = Offset(
-            xAxis.xFromEpoch(draggedPosition.dx.toInt()),
-            quoteToY(draggedPosition.dy)) +
+            xAxis.xFromEpoch(_draggedPosition.dx.toInt()),
+            quoteToY(_draggedPosition.dy)) +
         (isOtherEndDragged ? Offset.zero : delta);
 
-    draggedPosition = Offset(xAxis.epochFromX(localPosition.dx).toDouble(),
+    _draggedPosition = Offset(xAxis.epochFromX(localPosition.dx).toDouble(),
         quoteFromCanvasY(localPosition.dy));
   }
+
+  /// Creates a copy of this object.
+  DraggableEdgePoint copyWith({
+    int? epoch,
+    double? quote,
+    bool? isDrawingDragged,
+    bool? isDragged,
+  }) =>
+      DraggableEdgePoint(
+        epoch: epoch ?? this.epoch,
+        quote: quote ?? this.quote,
+        isDrawingDragged: isDrawingDragged ?? this.isDrawingDragged,
+        isDragged: isDragged ?? this.isDragged,
+      ).._draggedPosition = _draggedPosition;
 }
