@@ -1,4 +1,3 @@
-import 'package:collection/collection.dart';
 import 'package:deriv_chart/src/add_ons/add_ons_repository.dart';
 import 'package:deriv_chart/src/add_ons/drawing_tools_ui/drawing_tool_config.dart';
 import 'package:deriv_chart/src/add_ons/drawing_tools_ui/drawing_tools_dialog.dart';
@@ -8,10 +7,9 @@ import 'package:deriv_chart/src/deriv_chart/chart/chart.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/annotations/chart_annotation.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/chart_series/data_series.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/chart_series/series.dart';
-import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing.dart';
-import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_data.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/markers/marker_series.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/models/chart_object.dart';
+import 'package:deriv_chart/src/deriv_chart/drawing_tool_chart/drawing_tools.dart';
 import 'package:deriv_chart/src/misc/callbacks.dart';
 import 'package:deriv_chart/src/misc/chart_controller.dart';
 import 'package:deriv_chart/src/misc/extensions.dart';
@@ -92,11 +90,7 @@ class _DerivChartState extends State<DerivChart> {
   final AddOnsRepository<DrawingToolConfig> _drawingToolsRepo =
       AddOnsRepository<DrawingToolConfig>(DrawingToolConfig);
 
-  /// Selected drawing tool.
-  DrawingToolConfig? _selectedDrawingTool;
-
-  /// Existing drawings.
-  final List<DrawingData> _drawings = <DrawingData>[];
+  final DrawingTools _drawingTools = DrawingTools();
 
   @override
   void initState() {
@@ -174,10 +168,7 @@ class _DerivChartState extends State<DerivChart> {
                             ),
                           ))
                 ],
-                drawings: _drawings,
-                onAddDrawing: _onAddDrawing,
-                selectedDrawingTool: _selectedDrawingTool,
-                clearDrawingToolSelection: _clearDrawingToolSelection,
+                drawingTools: _drawingTools,
                 markerSeries: widget.markerSeries,
                 theme: widget.theme,
                 onCrosshairAppeared: widget.onCrosshairAppeared,
@@ -211,15 +202,11 @@ class _DerivChartState extends State<DerivChart> {
                 child: IconButton(
                   icon: const Icon(Icons.drive_file_rename_outline_outlined),
                   onPressed: () {
-                    /// Remove unfinished drawings before openning the dialog.
-                    /// For the scenario where the user adds part of a drawing
-                    /// and then opens the dialog.
                     setState(() {
-                      _drawings.removeWhere(
-                          (DrawingData data) => !data.isDrawingFinished);
-                      _selectedDrawingTool = null;
+                      _drawingTools
+                        ..init()
+                        ..drawingToolsRepo = _drawingToolsRepo;
                     });
-
                     showDialog<void>(
                       context: context,
                       builder: (
@@ -229,24 +216,7 @@ class _DerivChartState extends State<DerivChart> {
                               AddOnsRepository<DrawingToolConfig>>.value(
                         value: _drawingToolsRepo,
                         child: DrawingToolsDialog(
-                          onDrawingToolRemoval: (int index) {
-                            setState(() {
-                              _drawings.removeAt(index);
-                            });
-                          },
-                          onDrawingToolSelection:
-                              (DrawingToolConfig selectedDrawingTool) {
-                            setState(() {
-                              _selectedDrawingTool = selectedDrawingTool;
-                            });
-                          },
-                          onDrawingToolUpdate:
-                              (int index, DrawingToolConfig updatedConfig) {
-                            setState(() {
-                              _drawings[index] =
-                                  _drawings[index].updateConfig(updatedConfig);
-                            });
-                          },
+                          drawingTools: _drawingTools,
                         ),
                       ),
                     );
@@ -257,48 +227,4 @@ class _DerivChartState extends State<DerivChart> {
           ),
         ),
       );
-
-  void _onAddDrawing(
-    Map<String, List<Drawing>> addedDrawing, {
-    bool isDrawingFinished = false,
-  }) {
-    setState(() {
-      final String drawingId = addedDrawing.keys.first;
-
-      final DrawingData? existingDrawing = _drawings.firstWhereOrNull(
-        (DrawingData drawing) => drawing.id == drawingId,
-      );
-
-      if (existingDrawing == null) {
-        _drawings.add(DrawingData(
-          id: drawingId,
-          config: _selectedDrawingTool!,
-          drawingParts: addedDrawing.values.first,
-          isDrawingFinished: isDrawingFinished,
-        ));
-      } else {
-        existingDrawing
-          ..updateDrawingPartList(addedDrawing.values.first)
-          ..isSelected = true
-          ..isDrawingFinished = isDrawingFinished;
-      }
-
-      if (isDrawingFinished) {
-        _drawingToolsRepo.add(_selectedDrawingTool!);
-        _selectedDrawingTool = null;
-      }
-
-      if (_drawings.length > 1) {
-        _drawings.removeWhere((DrawingData data) =>
-            data.id != drawingId && !data.isDrawingFinished);
-      }
-    });
-  }
-
-  /// Clean the drawing tool selection.
-  void _clearDrawingToolSelection() {
-    setState(() {
-      _selectedDrawingTool = null;
-    });
-  }
 }
